Bring the workspace to zero clippy warnings and gate CI on it
162 warnings cleared: mechanical rewrites (From over Into, let-chains,
is_multiple_of, unsigned_abs, iterator loops, vec! init) plus Default
impls delegating to new(), and per-site allows with reasons where the
lint loses to the design — inline mapper RAM, frame-carrying message
enums, semantic Set-prefixed variants, the argument-heavy PPU signal
functions. CI's clippy step now runs with -D warnings, so any future
warning is a build failure rather than creep.

+++ crates/missingno-gb/src/audio/channels/wave.rs
@@ -300,7 +300,7 @@ impl WaveChannel {
         {
             self.frequency_timer -= 1;
             if self.frequency_timer == 0 {
-                self.frequency_timer = 2048 - self.period.0 as u16;
+                self.frequency_timer = 2048 - self.period.0;
                 self.pending_overflow = true;
             }
         }
@@ -341,7 +341,7 @@ impl WaveChannel {
         // the divider load window.
         self.wave_position = 0;
         self.output_dirty = true;
-        self.frequency_timer = 2048 - self.period.0 as u16;
+        self.frequency_timer = 2048 - self.period.0;
         self.ch3_frst = false;
         self.pending_overflow = false;
     }
